feat: extract extractCssFiles logic to extractCssFiles.js
